acme/requirements.txt
bsuite
pillow
tensorflow~=2.13.0
dm-control
trfl
optax==0.1.7
dm-sonnet
chex==0.1.7
rlds
jaxlib==0.4.14
absl-py
dm-haiku==0.0.10
atari-py
tensorflow_datasets
jax==0.4.14
typing-extensions
flax==0.7.4
rlax==0.1.6
gymnasium[box2d]~=0.26.0
pygame~=2.1.0
dm-env
numpy
gym[atari]~=0.25.0
dm-tree
dm-launchpad[tensorflow,reverb] @ https://github.com/ethanluoyc/launchpad/releases/download/v0.6.0rc0/dm_launchpad-0.6.0rc0-cp39-cp39-manylinux2014_x86_64.whl
tensorflow_probability>=0.17.0,<0.22
